refactor: move logging and cache infrastructure to infrastructure package
