+++ src/main/java/pl/cleankod/exchange/provider/nbp/ExchangeRatesNbpClientFactory.java
@@ -7,8 +7,8 @@ import feign.jackson.JacksonEncoder;
 import pl.cleankod.exchange.provider.nbp.decoder.ExchangeRatesNbpClienErrorDecoder;
 import pl.cleankod.exchange.provider.nbp.model.RateTableAndCurrencyCacheKey;
 import pl.cleankod.exchange.provider.nbp.model.RateWrapper;
-import pl.cleankod.util.cache.CacheInMemoryClient;
-import pl.cleankod.util.cache.inmemory.CacheInMemoryLru;
+import pl.cleankod.infrastructure.cache.CacheInMemoryClient;
+import pl.cleankod.infrastructure.cache.inmemory.CacheInMemoryLru;
 
 public class ExchangeRatesNbpClientFactory {
     public static ExchangeRatesNbpClient create(String baseUrl) {
